FIX: SequentialSubspaceMinimizer still used unsigned int for NormX and NormY.
- this causes buggy behavior when using floating norms.

+++ src/Minimizations/SequentialSubspaceMinimizer.cpp
@@ -22,8 +22,8 @@
 #include "MinimizationExceptions.hpp"
 
 SequentialSubspaceMinimizer::SequentialSubspaceMinimizer(
-		const unsigned int _NormX,
-		const unsigned int _NormY,
+		const double _NormX,
+		const double _NormY,
 		const double _PowerX,
 		const double _PowerY,
 		const double _Delta,
